[style] 코드 스타일 포맷 변경
구글 자바 포맷을 이용하여
코드를 2탭으로 수정하고, 라이브러리 임포트를 싱글 임포트로 변경해주었다.
