+++ .jpb/persistence-units.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="PersistenceUnitSettings">
+    <persistence-units>
+      <persistence-unit name="Default">
+        <packages>
+          <package value="f4.auth" />
+        </packages>
+      </persistence-unit>
+    </persistence-units>
+  </component>
+</project>
