+++ src/main/java/f4/auth/AuthApplication.java
@@ -6,8 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class AuthApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(AuthApplication.class, args);
-	}
-
+  public static void main(String[] args) {
+    SpringApplication.run(AuthApplication.class, args);
+  }
 }
