+++ src/main/java/f4/auth/domain/user/constant/Role.java
@@ -1,6 +1,6 @@
 package f4.auth.domain.user.constant;
 
 public enum Role {
-
-    USER, ADMIN;
+  USER,
+  ADMIN;
 }
