+++ src/main/java/f4/auth/domain/user/controller/AuthController.java
@@ -5,7 +5,12 @@ import f4.auth.domain.user.dto.response.TokenResponseDto;
 import f4.auth.domain.user.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 
@@ -14,17 +19,15 @@ import javax.validation.Valid;
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final AuthService authService;
+  private final AuthService authService;
 
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@Valid @RequestBody LoginRequestDto dto) {
-        return ResponseEntity.ok(authService.login(dto));
-    }
+  @PostMapping("/login")
+  public ResponseEntity<?> login(@Valid @RequestBody LoginRequestDto dto) {
+    return ResponseEntity.ok(authService.login(dto));
+  }
 
-
-    @PutMapping("/login")
-    public TokenResponseDto reIssue(@RequestHeader(value = "Refresh-Token") String rtk) {
-        return authService.reIssue(rtk);
-    }
-
+  @PutMapping("/login")
+  public TokenResponseDto reIssue(@RequestHeader(value = "Refresh-Token") String rtk) {
+    return authService.reIssue(rtk);
+  }
 }
