+++ src/main/java/f4/auth/domain/user/persist/repository/MemberRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
-    Optional<Member> findByEmail(String email);
+  Optional<Member> findByEmail(String email);
 }
