+++ src/main/java/f4/auth/domain/user/service/MemberService.java
@@ -4,5 +4,5 @@ import f4.auth.domain.user.dto.request.SignupRequestDto;
 
 public interface MemberService {
 
-    void register(SignupRequestDto signupRequestDto);
+  void register(SignupRequestDto signupRequestDto);
 }
