+++ src/main/java/f4/auth/global/exception/CustomException.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 @Getter
 public class CustomException extends RuntimeException {
 
-    private final CustomErrorCode customErrorCode;
+  private final CustomErrorCode customErrorCode;
 
-    public CustomException(CustomErrorCode customErrorCode) {
-        super(customErrorCode.getMessage());
-        this.customErrorCode = customErrorCode;
-    }
+  public CustomException(CustomErrorCode customErrorCode) {
+    super(customErrorCode.getMessage());
+    this.customErrorCode = customErrorCode;
+  }
 }
