+++ src/main/java/f4/auth/global/exception/ErrorDetails.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Builder
 public class ErrorDetails {
 
-    private String path;
-    private int code;
-    private String message;
+  private String path;
+  private int code;
+  private String message;
 }
