+++ src/main/java/f4/auth/global/security/jwt/exception/InvalidTokenException.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class InvalidTokenException extends CustomException {
 
-    public InvalidTokenException(CustomErrorCode customErrorCode) {
-        super(customErrorCode);
-    }
+  public InvalidTokenException(CustomErrorCode customErrorCode) {
+    super(customErrorCode);
+  }
 }
